@@ -27,81 +27,83 @@
   "homepage": "https://github.com/WATonomous/infra-config#readme",
   "description": "A website for WATcloud",
   "dependencies": {
-    "@giscus/react": "^3.0.0",
-    "@hookform/resolvers": "^3.3.3",
-    "@icons-pack/react-simple-icons": "^10.2.0",
-    "@radix-ui/react-alert-dialog": "^1.0.5",
-    "@radix-ui/react-avatar": "^1.1.1",
-    "@radix-ui/react-checkbox": "^1.0.4",
+    "@giscus/react": "^3.1.0",
+    "@hookform/resolvers": "^5.2.2",
+    "@icons-pack/react-simple-icons": "^13.8.0",
+    "@radix-ui/react-alert-dialog": "^1.1.15",
+    "@radix-ui/react-avatar": "^1.1.10",
+    "@radix-ui/react-checkbox": "^1.3.3",
     "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-popover": "^1.0.7",
-    "@radix-ui/react-scroll-area": "^1.0.5",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-separator": "^1.0.3",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-popover": "^1.1.15",
+    "@radix-ui/react-scroll-area": "^1.2.10",
+    "@radix-ui/react-select": "^2.2.6",
+    "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-switch": "^1.0.3",
-    "@radix-ui/react-tooltip": "^1.0.7",
-    "@rjsf/core": "^5.16.1",
-    "@rjsf/utils": "^5.16.1",
-    "@rjsf/validator-ajv8": "^5.16.1",
-    "@sentry/nextjs": "^7.76.0",
+    "@radix-ui/react-switch": "^1.2.6",
+    "@radix-ui/react-tooltip": "^1.2.8",
+    "@rjsf/core": "^5.24.13",
+    "@rjsf/utils": "^5.24.13",
+    "@rjsf/validator-ajv8": "^5.24.13",
+    "@sentry/nextjs": "^10.19.0",
     "@types/json-schema": "^7.0.15",
-    "@uppy/aws-s3": "^3.6.2",
-    "@uppy/dashboard": "^3.7.5",
-    "@uppy/react": "^3.2.2",
-    "bcrypt-ts": "^5.0.3",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
-    "cmdk": "^0.2.0",
-    "core-js": "^3.34.0",
-    "dayjs": "^1.11.10",
-    "js-base64": "^3.7.7",
-    "js-sha256": "^0.11.0",
-    "lucide-react": "^0.288.0",
-    "next": "^13.5.6",
-    "nextjs-google-analytics": "^2.3.3",
-    "nextra": "^2.13.2",
-    "nextra-theme-docs": "^2.13.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.49.2",
+    "@uppy/aws-s3": "^5.0.1",
+    "@uppy/core": "^5.1.0",
+    "@uppy/dashboard": "^5.0.2",
+    "@uppy/react": "^5.1.0",
+    "bcrypt-ts": "^7.1.0",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.1.1",
+    "core-js": "^3.46.0",
+    "dayjs": "^1.11.18",
+    "js-base64": "^3.7.8",
+    "js-sha256": "^0.11.1",
+    "lucide-react": "^0.545.0",
+    "next": "^13.5.11",
+    "next-seo": "^6.8.0",
+    "nextjs-google-analytics": "^2.3.7",
+    "nextra": "^3.3.1",
+    "nextra-theme-docs": "^3.3.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.65.0",
     "sha512crypt-node": "^1.0.2",
-    "sonner": "^1.4.41",
-    "strip-indent": "^4.0.0",
-    "swr": "^2.2.4",
+    "sonner": "^2.0.7",
+    "strip-indent": "^4.1.1",
+    "swr": "^2.3.6",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.22.4"
+    "zod": "^4.1.12"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^14.0.4",
-    "@types/jest": "^29.5.6",
-    "@types/node": "20.8.7",
-    "@types/react": "^18.2.31",
+    "@next/bundle-analyzer": "^14.2.33",
+    "@types/jest": "^30.0.0",
+    "@types/node": "24.7.2",
+    "@types/react": "^18.3.26",
     "ajv-cli": "^5.0.0",
-    "ajv-formats": "^2.1.1",
-    "autoprefixer": "^10.4.16",
-    "axios": "^1.9.0",
-    "dedent": "^1.5.3",
-    "eslint": "8.52.0",
-    "eslint-config-next": "13.5.6",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-mdx": "^2.2.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
+    "ajv-formats": "^3.0.1",
+    "autoprefixer": "^10.4.21",
+    "axios": "^1.12.2",
+    "dedent": "^1.7.0",
+    "eslint": "^9.37.0",
+    "eslint-config-next": "^15.5.5",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-mdx": "^3.6.2",
+    "eslint-plugin-react-hooks": "^7.0.0",
     "https-proxy-agent": "^7.0.6",
-    "jest": "^29.7.0",
-    "p-limit": "^6.2.0",
-    "postcss": "^8.4.31",
-    "quicktype": "^23.0.76",
-    "sharp-cli": "^4.2.0",
+    "jest": "^30.2.0",
+    "p-limit": "^7.1.1",
+    "postcss": "^8.5.6",
+    "quicktype": "^23.2.6",
+    "sharp-cli": "^5.2.0",
     "slugify": "^1.6.6",
-    "tailwindcss": "^3.4.10",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "tailwindcss": "^3.4.18",
+    "ts-jest": "^29.4.5",
+    "typescript": "^5.9.3"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.0.0"
   }
 }
