@@ -23,7 +23,7 @@ import { useRouter } from 'next/router';
 import { MdxFile } from "nextra";
 import { Link } from "nextra-theme-docs";
 import { getPagesUnderRoute } from "nextra/context";
-import { useState, useEffect } from "react";
+import { useState, useEffect, useMemo, useRef } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 import Picture from "./picture";
@@ -48,25 +48,38 @@ export function BlogIndex() {
     const router = useRouter();
     const locale = router.locale || websiteConfig.default_locale;
     const activeTag = (router.query.tag as string | undefined)?.trim();
-    let tagCounts: Record<string, number> = {};
+    const hasRedirectedEmptyTag = useRef(false);
+    const hasRedirectedInvalidTag = useRef(false);
+    const lastActiveTag = useRef(activeTag);
 
-    // Get all posts from route
-    const allPosts = getPagesUnderRoute("/blog").filter((page) => {
-        const frontMatter = (page as MdxFile).frontMatter || {};
-        // Get tag counts for the tag bar
-        if (frontMatter.tags && Array.isArray(frontMatter.tags)) {
-            frontMatter.tags.forEach((tag: string) => {
-                tagCounts[tag] = (tagCounts[tag] || 0) + 1;
-            });
-        }
-        if (frontMatter.hidden) {return null}
-        return frontMatter;
-    });
+    // Get all posts from route and calculate tag counts
+    const { allPosts, tagCounts } = useMemo(() => {
+        const tagCounts: Record<string, number> = {};
+        const allPosts = getPagesUnderRoute("/blog").filter((page) => {
+            const frontMatter = (page as MdxFile).frontMatter || {};
+            // Get tag counts for the tag bar
+            if (frontMatter.tags && Array.isArray(frontMatter.tags)) {
+                frontMatter.tags.forEach((tag: string) => {
+                    tagCounts[tag] = (tagCounts[tag] || 0) + 1;
+                });
+            }
+            if (frontMatter.hidden) {return null}
+            return frontMatter;
+        });
+        return { allPosts, tagCounts };
+    }, []);
     
     // Redirect to main blog page if no tag specified or empty tag
     // (but only after router is ready and we've attempted to parse the tag)
     useEffect(() => {
+        // Reset redirect flag when activeTag changes
+        if (activeTag !== lastActiveTag.current) {
+            hasRedirectedEmptyTag.current = false;
+            lastActiveTag.current = activeTag;
+        }
+
         if (router.isReady && !activeTag &&
+            !hasRedirectedEmptyTag.current &&
             (
                 !router.query.tag ||
                 (typeof router.query.tag === 'string' && router.query.tag.trim() === '') ||
@@ -75,18 +88,23 @@ export function BlogIndex() {
                 router.asPath.endsWith('?tag=')
             )
         ) {
-            router.push('/blog')
-        } // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [router.isReady, activeTag, router.query.tag, router.asPath])
-    // excluding router object from deps to prevent infinite loop (it changes on every navigation)
+            hasRedirectedEmptyTag.current = true;
+            router.push('/blog');
+        }
+    }, [router, activeTag])
 
     // Redirect if tag has no posts
     useEffect(() => {
-        if (router.isReady && activeTag && (!tagCounts[activeTag] || tagCounts[activeTag] === 0)) {
-            router.push('/blog')
-        } // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [router.isReady, activeTag])
-    // excluding router object from deps to prevent infinite loop (it changes on every navigation)
+        // Reset redirect flag when activeTag changes to a valid tag
+        if (activeTag && tagCounts[activeTag] && tagCounts[activeTag] > 0) {
+            hasRedirectedInvalidTag.current = false;
+        }
+
+        if (router.isReady && activeTag && !hasRedirectedInvalidTag.current && (!tagCounts[activeTag] || tagCounts[activeTag] === 0)) {
+            hasRedirectedInvalidTag.current = true;
+            router.push('/blog');
+        }
+    }, [router, activeTag, tagCounts])
 
     // Filter blogs by tag
     const filteredPosts = allPosts.filter((page) => {
@@ -219,8 +237,8 @@ export function BlogIndex() {
 }
 
 const subscribeFormSchema = z.object({
-    email: z.string().email({
-        message: "Please enter a valid email.",
+    email: z.email({
+        error: "Please enter a valid email.",
     }),
 });
 
