@@ -14,7 +14,7 @@ import { useState } from "react";
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 import { Input } from "./ui/input";
-import { useMDXComponents } from "nextra-theme-docs";
+import { useMDXComponents } from "nextra/mdx";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "./ui/card";
 import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar";
 import { Link } from "nextra-theme-docs"
@@ -217,7 +217,7 @@ export function UsernameToID() {
             {globalNodeID && (
                 <div className="mt-8">
                     <h2 className="text-xl font-bold">Global Node ID</h2>
-                    <Pre hasCopyCode><Code>{globalNodeID}</Code></Pre>
+                    <Pre><Code>{globalNodeID}</Code></Pre>
                 </div>
             )}
             {rawData && (
@@ -228,7 +228,7 @@ export function UsernameToID() {
                         <Details>
                             <Summary>Raw data</Summary>
                             <div className="mt-2">
-                                <Pre hasCopyCode><Code>{JSON.stringify(rawData, null, 2)}</Code></Pre>
+                                <Pre><Code>{JSON.stringify(rawData, null, 2)}</Code></Pre>
                             </div>
                         </Details>
                     </div>
