@@ -1,14 +1,14 @@
-import { useConfig } from "nextra-theme-docs"
+import { useTheme } from "next-themes"
 import { Toaster as Sonner } from "sonner"
 
 type ToasterProps = React.ComponentProps<typeof Sonner>
 
 const Toaster = ({ ...props }: ToasterProps) => {
-  const { darkMode } = useConfig();
+  const { resolvedTheme } = useTheme();
 
   return (
     <Sonner
-      theme={darkMode === true ? "dark" : darkMode === false ? "light" : "system"}
+      theme={resolvedTheme === "dark" ? "dark" : resolvedTheme === "light" ? "light" : "system"}
       className="toaster group"
       toastOptions={{
         classNames: {
