@@ -1,4 +1,4 @@
-{
+const metaConfig = {
   "index": {
     "title": "Welcome",
     "type": "page",
@@ -38,4 +38,6 @@
   "onboarding-form": {
     "display": "hidden"
   }
-}
+};
+
+export default metaConfig;
