@@ -1,8 +1,10 @@
-{
+const metaConfig = {
   "index": "Introduction",
   "services": "Services",
   "compute-cluster": "Compute Cluster",
   "registered-affiliations": "Registered Affiliations",
   "utilities": "Utilities",
   "community-docs": "Community Docs"
-}
+};
+
+export default metaConfig;
