@@ -1,7 +1,9 @@
-{
+const metaConfig =  {
   "*": {
     "theme": {
       "timestamp": false
     }
   }
-}
+};
+
+export default metaConfig;
