@@ -1,5 +1,7 @@
-{
+const metaConfig =  {
     "index": "Introduction",
     "sponsor": "Sponsor",
     "join": "Join WATcloud"
-}
+};
+
+export default metaConfig;
